pyproject.toml
[project]
name = "Expirements"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "tqdm>=4.67.1",
]
